WIP: Throw errors when applying deltas that depend on missing state
Introduces a new "delta depends on missing vertex" error.

Previously, applying deltas would succeed, but result in inconsistent
results when state was applied afterwards.

For the 'push' and 'insertAt' mutations, this removes the redundant
initial added vertex (but leaves the edge) ... and uses that as a
signal to check to ensure that the vertex exists when applying the
delta.

TODO: Other mutations: addRight, remove, removeAt